# optimize Adam and Adamax optimizer