--- a/spark/dl/src/main/scala/com/intel/analytics/bigdl/optim/Adam.scala
+++ b/spark/dl/src/main/scala/com/intel/analytics/bigdl/optim/Adam.scala
@@ -17,7 +17,7 @@
 package com.intel.analytics.bigdl.optim
 
 import breeze.linalg.*
-import com.intel.analytics.bigdl.tensor.Tensor
+import com.intel.analytics.bigdl.tensor.{FloatType, Tensor}
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.{T, Table}
 
@@ -39,6 +39,17 @@ class Adam[@specialized(Float, Double) T: ClassTag](
  var beta1: Double = 0.9,
  var beta2: Double = 0.999,
  var Epsilon: Double = 1e-8)(implicit ev: TensorNumeric[T]) extends OptimMethod[T] {
+
+  @transient
+  val buffer1 = Tensor[T]()
+  @transient
+  val buffer2 = Tensor[T]()
+  @transient
+  val buffer3 = Tensor[T]()
+  @transient
+  val buffer4 = Tensor[T]()
+  @transient
+  val buffer5 = Tensor[T]()
 
   /**
    * An implementation of Adam http://arxiv.org/pdf/1412.6980.pdf
@@ -64,22 +75,55 @@ class Adam[@specialized(Float, Double) T: ClassTag](
     val (_s, _r, _denom) =
       if (state.get[Tensor[T]]("s").isDefined) {
         (state.get[Tensor[T]]("s").get, state.get[Tensor[T]]("r").get,
-          Tensor[T]().resizeAs(dfdx).zero())
+          buffer3.resizeAs(dfdx))
       } else {
-        (Tensor[T]().resizeAs(dfdx).zero(), Tensor[T]().resizeAs(dfdx).zero(),
-          Tensor[T]().resizeAs(dfdx).zero())
+        (buffer1.resizeAs(dfdx).zero(), buffer2.resizeAs(dfdx).zero(),
+          buffer3.resizeAs(dfdx))
       }
+    buffer4.resizeAs(dfdx)
+
+    // used as MKL.axpy: 1 * a + y = y
+    if (buffer5.nElement < dfdx.nElement) {
+      buffer5.resizeAs(dfdx).fill(ev.fromType[Int](1))
+    }
+
     val clr = lr / (1 + timestep*lrd)
 
     timestep = timestep + 1
 
+    /**
+     * m_t = beta_1 * m_t-1 + (1 - beta_1) * g_t
+     * v_t = beta_2 * v_t-1 + (1 - beta_2) * g_t * g_t
+     */
     _s.mul(ev.fromType[Double](beta1)).add(ev.fromType[Double](1-beta1), dfdx)
-    _r.mul(ev.fromType[Double](beta2)).addcmul(ev.fromType[Double](1-beta2), dfdx, dfdx)
-    _denom.resizeAs(_r).copy(_r).sqrt().add(ev.fromType[Double](eps))
+    ev.vMul(dfdx.nElement,
+      dfdx.storage.array,
+      dfdx.storageOffset - 1,
+      dfdx.storage.array,
+      dfdx.storageOffset - 1,
+      buffer4.storage.array,
+      buffer4.storageOffset - 1)
+    _r.mul(ev.fromType[Double](beta2)).add(ev.fromType[Double](1-beta2), buffer4)
+
+    ev.vSqrt(_r.nElement,
+      _r.storage.array,
+      _r.storageOffset - 1,
+      _denom.storage.array,
+      _denom.storageOffset - 1)
+
+    ev.axpy(_denom.nElement,
+      ev.fromType[Double](eps),
+      buffer5.storage.array,
+      buffer5.storageOffset - 1,
+      1,
+      _denom.storage.array,
+      _denom.storageOffset - 1,
+      1)
     // efficiency improved upon by changing the order of computation, at expense of clarity
     val biasCorrection1 = 1 - pow(beta1, timestep)
     val biasCorrection2 = 1 - pow(beta2, timestep)
     val stepSize = clr * sqrt(biasCorrection2) / biasCorrection1
+
     parameter.addcdiv(ev.fromType[Double](-stepSize), _s, _denom)
 
     state("evalCounter") = timestep // A tmp tensor to hold the sqrt(v) + epsilon