@jason-dai 2017-08-07
is the test single-threading or multi-threading?